backend/requirements.txt
fastapi[all]
uvicorn
sqlalchemy
psycopg2-binary
cryptography
pydantic
pyjwt